+++ src/crawlee/storage_clients/_base/_dataset_client.py
@@ -16,8 +16,8 @@ if TYPE_CHECKING:
 
 
 @docs_group('Abstract classes')
-class BaseDatasetClient(ABC):
-    """Abstract base class for dataset resource clients.
+class DatasetClient(ABC):
+    """An abstract class for dataset resource clients.
 
     These clients are specific to the type of resource they manage and operate under a designated storage
     client, like a memory storage client.